qlearning-agent/requirements.txt
pyzmq
numpy
scipy
tensorflow==2.11.0
scikit-learn
